+++ WhistleWindLobotomyMod/Cards/Lobotomy Corporation/youreBald.cs
@@ -1,5 +1,6 @@
 ﻿using DiskCardGame;
 using InscryptionAPI.Card;
+using WhistleWind.AbnormalSigils;
 using WhistleWind.Core.Helpers;
 
 using static WhistleWindLobotomyMod.Core.LobotomyCardManager;
@@ -10,11 +11,11 @@ namespace WhistleWindLobotomyMod {
         private static void YoureBald_BaldIsAwesome() {
             string textureName = "youreBald";
             CardManager.New(LobotomyPlugin.pluginPrefix, youreBald, "You're Bald...",
-                attack: 0, health: 2, "I've always wondered what it's like to be bald.")
+                attack: 0, health: 0, "I've always wondered what it's like to be bald.")
                 .SetEnergyCost(2)
                 .SetPortraits(LobotomyPlugin.ModAssembly, textureName)
-                .AddAbilities(Ability.DrawCopy)
-                .SetDefaultEvolutionName("You're Really Bald...")
+                .AddAbilities(Shaver.ability)
+                .SetSpellType(SpellType.Targeted)
                 .Build(CardHelper.CardType.Common, RiskLevel.Zayin, true);
         }
     }